requirements.txt
# Pipeline ETL & base de données
pandas>=2.0
openpyxl>=3.1
sqlalchemy>=2.0
psycopg2-binary>=2.9
python-dotenv>=1.0

# Dashboard
streamlit>=1.35
plotly>=5.20

# Statistiques
statsmodels>=0.14

# Export PDF
fpdf2>=2.7

# Authentification & ML
streamlit-authenticator>=0.3
scikit-learn>=1.4

# Tests
pytest>=8.0
pytest-cov>=5.0
